# docs/swagger.yaml
components:
  schemas:
    CasTask:
      properties:
        task_id:
          format: uuid4
          title: Task Id
          type: string
        task_status:
          title: Task Status
          type: string
      required:
      - task_id
      - task_status
      title: CasTask
      type: object
    CustomersForAllCategoriesAnalysis:
      properties:
        analysis_value_avg:
          description: Average value of the analysis for the customer across all reviews
            / comments
          title: Analysis Value Avg
          type: number
        customers_count:
          description: Count of customers of the product
          minimum: 0.0
          title: Customers Count
          type: integer
        href_category_1:
          description: Href to the first level category
          title: Href Category 1
          type: string
        href_category_2:
          description: Href to the second level category
          title: Href Category 2
          type: string
        href_category_3:
          description: Href to the third level category
          title: Href Category 3
          type: string
        href_category_4:
          description: Href to the fourth level category
          title: Href Category 4
          type: string
        product_fullname:
          description: The full name of the analysis product
          title: Product Fullname
          type: string
        ru_category_1:
          description: Title of the first level category in russian
          title: Ru Category 1
          type: string
        ru_category_2:
          description: Title of the second level category in russian
          title: Ru Category 2
          type: string
        ru_category_3:
          description: Title of the third level category in russian
          title: Ru Category 3
          type: string
        ru_category_4:
          description: Title of the fourth level category in russian
          title: Ru Category 4
          type: string
      required:
      - ru_category_1
      - href_category_1
      - ru_category_2
      - href_category_2
      - ru_category_3
      - href_category_3
      - ru_category_4
      - href_category_4
      - product_fullname
      - customers_count
      - analysis_value_avg
      title: Analysis value customers across all categories
      type: object
...
    GroupRegionallyAllCustomerAnalysis:
      properties:
        analysis_value_avg:
          description: Average value of the analysis for the customer across all reviews
            / comments
          title: Analysis Value Avg
          type: number
        city_en:
          description: Customer city in english
          title: City En
          type: string
        city_ru:
          description: Customer city in russian
          title: City Ru
          type: string
        country_en:
          description: Customer country in english
          title: Country En
          type: string
        country_ru:
          description: Customer country in russian
          title: Country Ru
          type: string
        latitude:
          description: Latitude of the customer region
          title: Latitude
          type: number
        longitude:
          description: Longitude of the customer region
          title: Longitude
          type: number
        object_count:
          description: Count of reviews / comments on the analysis by region
          minimum: 0.0
          title: Object Count
          type: integer
      required:
      - country_ru
      - country_en
      - city_ru
      - city_en
      - latitude
      - longitude
      - object_count
      - analysis_value_avg
      title: Customer analysis values by region
      type: object
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
    VisualizationType:
      description: An enumeration.
      enum:
      - image
      - html
      title: VisualizationType
      type: string
info:
  title: Customer Analysis Service API
  version: 1.0.0
openapi: 3.0.2
paths:
...
  /api/v1/visualizer/analysis_value/category:
    post:
      description: "\u0412\u0438\u0437\u0443\u0430\u043B\u0438\u0437\u0438\u0440\u043E\
        \u0432\u0430\u0442\u044C \u0437\u043D\u0430\u0447\u0435\u043D\u0438\u044F\
        \ \u0430\u043D\u0430\u043B\u0438\u0437\u0430, \u0433\u0440\u0443\u043F\u043F\
        \u0438\u0440\u0443\u044F \u043F\u043E \u0432\u0441\u0435\u043C \u043A\u0430\
        \u0442\u0435\u0433\u043E\u0440\u0438\u044F\u043C. (\u0434\u043B\u044F analysis_sentiments\
        \ category)"
      operationId: visualize_analysis_value_by_groupings_by_category_api_v1_visualizer_analysis_value_category_post
      parameters:
      - in: query
        name: vis_type
        required: true
        schema:
          $ref: '#/components/schemas/VisualizationType'
      - description: "\u0417\u0430\u0433\u043E\u043B\u043E\u0432\u043E\u043A \u0432\
          \u0438\u0437\u0430\u0443\u043B\u0438\u0437\u0430\u0446\u0438\u0438 \u0430\
          \u043D\u0430\u043B\u0438\u0442\u0438\u043A\u0438"
        in: query
        name: title
        required: false
        schema:
          default: Customer Analysis
          description: "\u0417\u0430\u0433\u043E\u043B\u043E\u0432\u043E\u043A \u0432\
            \u0438\u0437\u0430\u0443\u043B\u0438\u0437\u0430\u0446\u0438\u0438 \u0430\
            \u043D\u0430\u043B\u0438\u0442\u0438\u043A\u0438"
          maxLength: 100
          minLength: 3
          title: Title
          type: string
      - description: "\u041E\u0431\u043E\u0437\u043D\u0430\u0447\u0435\u043D\u0438\
          \u0435 \u043A\u043E\u043B\u0438\u0447\u0435\u0441\u0442\u0432\u0435\u043D\
          \u043D\u043E\u0433\u043E \u0437\u043D\u0430\u0447\u0435\u043D\u0438\u044F\
          \ \u043A\u043B\u0438\u0435\u043D\u0442\u044B / \u043E\u0442\u0437\u044B\u0432\
          \u044B / \u043A\u043E\u043C\u043C\u0435\u043D\u0442\u0430\u0440\u0438\u0438"
        in: query
        name: title_object_count
        required: false
        schema:
          default: Count customers
          description: "\u041E\u0431\u043E\u0437\u043D\u0430\u0447\u0435\u043D\u0438\
            \u0435 \u043A\u043E\u043B\u0438\u0447\u0435\u0441\u0442\u0432\u0435\u043D\
            \u043D\u043E\u0433\u043E \u0437\u043D\u0430\u0447\u0435\u043D\u0438\u044F\
            \ \u043A\u043B\u0438\u0435\u043D\u0442\u044B / \u043E\u0442\u0437\u044B\
            \u0432\u044B / \u043A\u043E\u043C\u043C\u0435\u043D\u0442\u0430\u0440\u0438\
            \u0438"
          maxLength: 50
          minLength: 3
          title: Title Object Count
          type: string
      - description: "\u041E\u0431\u043E\u0437\u043D\u0430\u0447\u0435\u043D\u0438\
          \u0435 \u0437\u043D\u0430\u0447\u0435\u043D\u0438\u044F \u0430\u043D\u0430\
          \u043B\u0438\u0437\u0430 \u043D\u0430\u0441\u0442\u0440\u043E\u0435\u043D\
          \u0438\u0435 / \u0441\u0445\u043E\u0436\u0435\u0441\u0442\u044C / \u0442\
          .\u043F."
        in: query
        name: title_analysis_value
        required: false
        schema:
          default: Analysis value
          description: "\u041E\u0431\u043E\u0437\u043D\u0430\u0447\u0435\u043D\u0438\
            \u0435 \u0437\u043D\u0430\u0447\u0435\u043D\u0438\u044F \u0430\u043D\u0430\
            \u043B\u0438\u0437\u0430 \u043D\u0430\u0441\u0442\u0440\u043E\u0435\u043D\
            \u0438\u0435 / \u0441\u0445\u043E\u0436\u0435\u0441\u0442\u044C / \u0442\
            .\u043F."
          maxLength: 50
          minLength: 3
          title: Title Analysis Value
          type: string
      requestBody:
        content:
          application/json:
            schema:
              items:
                $ref: '#/components/schemas/CustomersForAllCategoriesAnalysis'
              title: Data
              type: array
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CasTask'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Visualize analysis values by grouping by category
      tags:
      - visualizer_analysis_value
  /api/v1/visualizer/analysis_value/region:
    post:
      description: "\u0412\u0438\u0437\u0443\u0430\u043B\u0438\u0437\u0438\u0440\u043E\
        \u0432\u0430\u0442\u044C \u0437\u043D\u0430\u0447\u0435\u043D\u0438\u044F\
        \ \u0430\u043D\u0430\u043B\u0438\u0437\u0430, \u0433\u0440\u0443\u043F\u043F\
        \u0438\u0440\u0443\u044F \u043F\u043E \u0432\u0441\u0435\u043C \u0440\u0435\
        \u0433\u0438\u043E\u043D\u0430\u043C \u043A\u043B\u0438\u0435\u043D\u0442\u043E\
        \u0432. (\u0434\u043B\u044F analysis_sentiments region)"
      operationId: visualize_analysis_value_by_groupings_by_regions_api_v1_visualizer_analysis_value_region_post
      parameters:
      - in: query
        name: vis_type
        required: true
        schema:
          $ref: '#/components/schemas/VisualizationType'
      - description: "\u0417\u0430\u0433\u043E\u043B\u043E\u0432\u043E\u043A \u0432\
          \u0438\u0437\u0430\u0443\u043B\u0438\u0437\u0430\u0446\u0438\u0438 \u0430\
          \u043D\u0430\u043B\u0438\u0442\u0438\u043A\u0438"
        in: query
        name: title
        required: false
        schema:
          default: Customer Analysis
          description: "\u0417\u0430\u0433\u043E\u043B\u043E\u0432\u043E\u043A \u0432\
            \u0438\u0437\u0430\u0443\u043B\u0438\u0437\u0430\u0446\u0438\u0438 \u0430\
            \u043D\u0430\u043B\u0438\u0442\u0438\u043A\u0438"
          maxLength: 100
          minLength: 3
          title: Title
          type: string
      - description: "\u041E\u0431\u043E\u0437\u043D\u0430\u0447\u0435\u043D\u0438\
          \u0435 \u043A\u043E\u043B\u0438\u0447\u0435\u0441\u0442\u0432\u0430 \u043E\
          \u0431\u044A\u0435\u043A\u0442\u043E\u0432 \u0430\u043D\u0430\u043B\u0438\
          \u0437\u0430 \u043E\u0442\u0437\u044B\u0432\u044B / \u043A\u043E\u043C\u043C\
          \u0435\u043D\u0442\u0430\u0440\u0438\u0438"
        in: query
        name: title_object_count
        required: false
        schema:
          default: Count reviews
          description: "\u041E\u0431\u043E\u0437\u043D\u0430\u0447\u0435\u043D\u0438\
            \u0435 \u043A\u043E\u043B\u0438\u0447\u0435\u0441\u0442\u0432\u0430 \u043E\
            \u0431\u044A\u0435\u043A\u0442\u043E\u0432 \u0430\u043D\u0430\u043B\u0438\
            \u0437\u0430 \u043E\u0442\u0437\u044B\u0432\u044B / \u043A\u043E\u043C\
            \u043C\u0435\u043D\u0442\u0430\u0440\u0438\u0438"
          maxLength: 50
          minLength: 3
          title: Title Object Count
          type: string
      - description: "\u041E\u0431\u043E\u0437\u043D\u0430\u0447\u0435\u043D\u0438\
          \u0435 \u0437\u043D\u0430\u0447\u0435\u043D\u0438\u044F \u0430\u043D\u0430\
          \u043B\u0438\u0437\u0430 \u043D\u0430\u0441\u0442\u0440\u043E\u0435\u043D\
          \u0438\u0435 / \u0441\u0445\u043E\u0436\u0435\u0441\u0442\u044C / \u0442\
          .\u043F."
        in: query
        name: title_analysis_value
        required: false
        schema:
          default: Analysis value
          description: "\u041E\u0431\u043E\u0437\u043D\u0430\u0447\u0435\u043D\u0438\
            \u0435 \u0437\u043D\u0430\u0447\u0435\u043D\u0438\u044F \u0430\u043D\u0430\
            \u043B\u0438\u0437\u0430 \u043D\u0430\u0441\u0442\u0440\u043E\u0435\u043D\
            \u0438\u0435 / \u0441\u0445\u043E\u0436\u0435\u0441\u0442\u044C / \u0442\
            .\u043F."
          maxLength: 50
          minLength: 3
          title: Title Analysis Value
          type: string
      requestBody:
        content:
          application/json:
            schema:
              items:
                $ref: '#/components/schemas/GroupRegionallyAllCustomerAnalysis'
              title: Data
              type: array
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CasTask'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Visualize analysis values by grouping by regions
      tags:
      - visualizer_analysis_value
...